課題3修正
